feat(frontend): tabbed Results page
Add tabs to the Results page, so that we can save multiple sets of results.
